+++ src/common-select/common-multi-select.jsx
@@ -21,6 +21,8 @@ type CommonMultiSelectProps = {
   initialOptions?: Array<CommonSelectOption>,
   /** Renders the input at its full width */
   isFullWidth?: boolean,
+  /** Label for the select. */
+  label?: string,
   /** Text displayed directly under the input */
   helperText?: string,
   /**
@@ -29,6 +31,8 @@ type CommonMultiSelectProps = {
    *  - `root`: card root div element
    *
    *  - `input`: input component
+   *
+   *  - `label`: input label component
    *
    *  - `options`: option dropdown items
    */