+++ src/common-select/common-select-component.jsx
@@ -3,8 +3,10 @@ import "./common-select.scss";
 import FormControl from "@material-ui/core/FormControl";
 import FormHelperText from "@material-ui/core/FormHelperText";
 import Input from "@material-ui/core/Input";
-import React from "react";
+import InputLabel from "@material-ui/core/InputLabel";
+import React, { useState } from "react";
 import classNames from "classnames";
+import isEmpty from "lodash/isEmpty";
 import {
   ClearIndicator,
   CommonSelectContainer,
@@ -16,6 +18,7 @@ import {
 } from "./components";
 
 type Props = {
+  value?: CommonSelectOption | Array<CommonSelectOption> | null,
   classes?: CommonSelectClasses,
   isFullWidth?: boolean,
   helperText?: string,
@@ -26,6 +29,7 @@ type Props = {
   isDisabled?: boolean,
   menuIsOpen?: boolean,
   onChange: Function,
+  label?: string,
   placeholder?: string,
   selectType?: "simple" | "async" | "creatable",
   showError?: boolean,
@@ -45,16 +49,30 @@ export const CommonSelectComponent = (props: Props) => {
     initialMessage = "",
     isClearable = true,
     isDisabled = false,
+    label = "",
     placeholder = "",
     showError = false,
     ...otherProps
   } = props;
+  const { value } = otherProps;
+  const [isInputFocused, setIsInputFocused] = useState(false);
+
   return (
     <div
       data-testid="common-select"
       className={classNames("common-select__root", classes.root)}
     >
       <FormControl fullWidth={isFullWidth}>
+        {label && (
+          <InputLabel
+            data-testid="common-select-input-label"
+            classes={{ formControl: classNames("common-select__input-label", classes.label) }}
+            htmlFor={id}
+            shrink={isInputFocused || !isEmpty(value)}
+          >
+            {label}
+          </InputLabel>
+        )}
         <Input
           inputComponent={CommonSelectContainer}
           disabled={isDisabled}
@@ -87,6 +105,8 @@ export const CommonSelectComponent = (props: Props) => {
               MultiValueRemove,
             },
           }}
+          onFocus={setIsInputFocused.bind(this, true)}
+          onBlur={setIsInputFocused.bind(this, false)}
           classes={{
             root: classNames("common-select__input-root", classes.input),
           }}